+++ rpcs3/Emu/RSX/GL/GLShaderInterpreter.cpp
@@ -1,0 +1,7 @@
+#include "stdafx.h"
+#include "GLShaderInterpreter.h"
+
+namespace gl
+{
+
+}
