+++ rpcs3/Emu/RSX/VK/VKShaderInterpreter.cpp
@@ -1,0 +1,7 @@
+#include "stdafx.h"
+#include "VKShaderInterpreter.h"
+
+namespace vk
+{
+
+};
