+++ rpcs3/Emu/RSX/VK/VKShaderInterpreter.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "VKGSRender.h"
+
+namespace vk
+{
+	class shader_interpreter : glsl::program
+	{
+	};
+}
